+++ FlashCard/06_FRONTEND/src/app/components/deck-flashcards/deck-flashcards.component.ts
@@ -5,6 +5,8 @@ import { Router, ActivatedRoute } from "@angular/router";
 import { RegisterService } from "../../services/auth/user.service";
 import { VoteService } from "../../services/vote/vote.service";
 import { MatSnackBar } from "@angular/material/snack-bar";
+import { MatDialog } from "@angular/material/dialog";
+import { ConfirmationDialogComponent } from "../confirmation-dialog/confirmation-dialog.component";
 
 @Component({
   selector: "app-deck-flashcards",
@@ -21,7 +23,8 @@ export class DeckFlashcardsComponent implements OnInit {
     private route: ActivatedRoute,
     private userService: RegisterService,
     private voteService: VoteService,
-    private snackBar: MatSnackBar
+    private snackBar: MatSnackBar,
+    private dialog : MatDialog
   ) {}
 
   ngOnInit(): void {
@@ -127,21 +130,33 @@ export class DeckFlashcardsComponent implements OnInit {
 
   // Method to delete a flashcard
   deleteFlashcard(flashcardId: string) {
-    // Call the deleteDeck() API method
-    this.flashcardService.deleteFlashcard(flashcardId).then(
-      (res) => {
-        console.log("Flashcard deleted successfully:", res);
-        // Reload decks after deletion
-        if (this.deckId) {
-          this.fetchFlashcards(this.deckId);
-        }
-        this.snackBar.open("Flashcard deleted successfully", "", { duration: 3000 });
-      },
-      (error) => {
-        // Handle error
-        console.error("Failed to delete Flashcard :", error);
-        this.snackBar.open("Failed to delete Flashcard. Please try again.", "", { duration: 3000 }); 
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      width: '250px',
+      data: { title: 'Confirm Deletion', message: 'Are you sure you want to delete this flashcard?' }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        // User confirmed deletion
+        this.flashcardService.deleteFlashcard(flashcardId).then(
+          (res) => {
+            this.snackBar.open("Flashcard deleted successfully", "", {
+              duration: 3000,
+            });
+            console.log("Flashcard deleted successfully:", res);
+            // Reload flashcards after deletion
+            if(this.deckId)
+              this.fetchFlashcards(this.deckId);
+          },
+          (error) => {
+            // Handle error
+            console.error("Failed to delete Flashcard :", error);
+            this.snackBar.open("Failed to delete Flashcard", "", {
+              duration: 3000,
+            });
+          }
+        );
       }
-    );
+    });
   }
 }
